Restructure: Moved routes to root, file/function rename
